pyproject.toml
# Copyright 2024 Telespazio UK Ltd
#    
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "data-transfer-events"
version = "0.1.0"
description = "Data transfer accounting events for transfers from workspace stores"
readme = "README.md"

# Our target version is Python 3.12, but it may work with earlier versions.
requires-python = ">=3.12"

license = {file = "LICENSE"}

# This should be your name or the name of the organization who originally
# authored the project, and a valid email address corresponding to the name
# listed.
authors = [
  {name = "Telespazio UK Ltd", email = "alex.hayward@telespazio.com" }
]

maintainers = [
  {name = "Telespazio UK Ltd", email = "shahid.raza@telespazio.com" }
]

# Classifiers help users find your project by categorizing it.
#
# For a list of valid classifiers, see https://pypi.org/classifiers/
classifiers = [
  # How mature is this project? Common values are
  #   3 - Alpha
  #   4 - Beta
  #   5 - Production/Stable
  "Development Status :: 3 - Alpha",

  # Indicate who your project is intended for
  "Intended Audience :: Science/Research",
  "Intended Audience :: Developers",

  # CHANGEME Pick your license as you wish
  "License :: OSI Approved :: Apache Software License",
  # "License :: Other/Proprietary License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3 :: Only",
]

# This field lists other packages that your project depends on to run.
# Any package you put here will be installed by pip when your project is
# installed, so they must be valid existing projects.
#
# For an analysis of this field vs pip's requirements files see:
# https://packaging.python.org/discussions/install-requires-vs-requirements/
dependencies = [
  "boto3",
  "pulsar-client",
  "pysubnettree",
  "requests",
  "eodhp-utils@git+https://github.com/UKEODHP/eodhp-utils@v0.1.6",
]

# List additional groups of dependencies here (e.g. development
# dependencies). Users will be able to install these using the "extras"
# syntax, for example:
#
#   $ pip install sampleproject[dev]
#
# Similar to `dependencies` above, these must be valid existing
# projects.
[project.optional-dependencies] # Optional
dev = [
    "pip-tools",
    "pytest", 
    "pytest-xdist", 
    "pytest-mock", 
    "pytest-watcher",
    "httpx",
    "validate-pyproject[all]",
    "pre-commit",
    "black",
    "moto",
    "isort",
    "ruff"
]

# List URLs that are relevant to your project
#
# This field corresponds to the "Project-URL" and "Home-Page" metadata fields:
# https://packaging.python.org/specifications/core-metadata/#project-url-multiple-use
# https://packaging.python.org/specifications/core-metadata/#home-page-optional
#
# Examples listed include a pattern for specifying where the package tracks
# issues, where the source is hosted, where to say thanks to the package
# maintainers, and where to support the project financially. The key is
# what's used to render the link text on PyPI.
[project.urls]  # Optional
homepage = "https://github.com/EO-DataHub/"
repository = "https://github.com/EO-DataHub/eodhp-data-transfer-events"
changelog = "https://github.com/EO-DataHub/eodhp-data-transfer-events/blob/main/CHANGELOG.md"

[build-system]
# These are the assumed default build requirements from pip:
# https://pip.pypa.io/en/stable/reference/pip/#pep-517-and-518-support
requires = ["setuptools>=69.0.0", "wheel", "setuptools-git-versioning>=2.0,<3"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
  "integrationtest: Integration test"
]

[tool.setuptools]
packages = []

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.pylint.'MESSAGES CONTROL']
disable = '''
    line-too-long,
    missing-class-docstring, 
    too-many-locals, 
    too-many-instance-attributes,
    logging-fstring-interpolation,
'''

[tool.isort]
profile = "black"